# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PuntosFieles API",
    "version": "1.3.7",
    "description": "Route-derived OpenAPI document for versioned /api/v1 routes, with curated request/response metadata layered onto the discovered route surface."
  },
  "servers": [
    {
      "url": "https://api.puntos.local"
    }
  ],
  "paths": {
...
    "/api/v1/admin/payment-webhooks": {
      "get": {
        "summary": "Derived GET /api/v1/admin/payment-webhooks",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentWebhookListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER",
          "MANAGER"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 30
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ]
      }
    },
    "/api/v1/admin/payment-webhooks/{id}/resolve": {
      "post": {
        "summary": "Derived POST /api/v1/admin/payment-webhooks/{id}/resolve",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentWebhookResolveResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER",
          "MANAGER"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentWebhookResolveRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/payments/webhook/{provider}": {
      "post": {
        "summary": "Derived POST /api/v1/public/payments/webhook/{provider}",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_staff"
      },
      "customerAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_customer"
      },
      "superAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_super"
      }
    },
    "schemas": {
...
      "PaymentWebhookResolveRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "customerPhone": {
            "type": "string",
            "minLength": 6
          }
        },
        "additionalProperties": false
      },
...
      "PaymentWebhookListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "provider": {
                  "type": "string"
                },
                "provider_event_id": {
                  "type": "string"
                },
                "event_type": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "business_slug": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "business_id": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uuid"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "customer_id": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uuid"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "customer_phone": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "amount_q": {
                  "type": [
                    "number",
                    "string"
                  ]
                },
                "currency": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "status": {
                  "type": "string"
                },
                "reason": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "error": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "linked_transaction_id": {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "uuid"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "created_at": {
                  "type": "string"
                },
                "updated_at": {
                  "$ref": "#/definitions/PaymentWebhookListResponse/properties/events/items/properties/created_at"
                }
              },
              "required": [
                "id",
                "provider",
                "provider_event_id",
                "amount_q",
                "status"
              ],
              "additionalProperties": true
            }
          }
        },
        "required": [
          "ok",
          "events"
        ],
        "additionalProperties": false
      },
      "PaymentWebhookResolveResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "event": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "provider": {
                "type": "string"
              },
              "provider_event_id": {
                "type": "string"
              },
              "event_type": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "business_slug": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "business_id": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "uuid"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "customer_id": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "uuid"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "customer_phone": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "amount_q": {
                "type": [
                  "number",
                  "string"
                ]
              },
              "currency": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "status": {
                "type": "string"
              },
              "reason": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "error": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "linked_transaction_id": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "uuid"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "created_at": {
                "type": "string"
              },
              "updated_at": {
                "$ref": "#/definitions/PaymentWebhookResolveResponse/properties/event/properties/created_at"
              }
            },
            "required": [
              "id",
              "provider",
              "provider_event_id",
              "amount_q",
              "status"
            ],
            "additionalProperties": true
          },
          "transaction": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": true
      },
...
    }
  }
}